fetched categories from db
