+++ frontend/src/App.js
@@ -1,5 +1,6 @@
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import ModalContextProvider from './contexts/ModalContextProvider';
+import CategoryProvider from './contexts/CategoryContext';
 import Home from './pages/Home';
 import LoginModal from './components/modals/LoginModal';
 import RegisterModal from './components/modals/RegisterModal';
@@ -9,7 +10,6 @@ import Catagories from './pages/Catagories';
 import MyProductLists from './pages/MyProductLists';
 import MyProfile from './pages/MyProfile';
 import Page404 from './pages/Page404';
-import CategoryContextProvider from './contexts/CategoryContextProvider';
 import CategoryModal from './components/modals/CategoryModal';
 
 
@@ -17,7 +17,7 @@ function App() {
   return (
     <div className="App">
       <ModalContextProvider>
-        <CategoryContextProvider>
+        <CategoryProvider>
       <Router>
         <header className="App-header">
           <Navbar />
@@ -39,7 +39,7 @@ function App() {
           </main>
           <footer></footer>
         </Router>
-        </CategoryContextProvider>
+        </CategoryProvider>
       </ModalContextProvider>
     </div>
   );
